+++ src/main/java/com/gerstox/projects/motostore_backend/repositories/LicenseRepository.java
@@ -3,4 +3,4 @@ package com.gerstox.projects.motostore_backend.repositories;
 import com.gerstox.projects.motostore_backend.entities.License;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LicenseRepository extends JpaRepository<License, Integer> {}
+public interface LicenseRepository extends JpaRepository<License, Long> {}
