+++ src/main/java/com/gerstox/projects/motostore_backend/repositories/RechargeRepository.java
@@ -3,4 +3,4 @@ package com.gerstox.projects.motostore_backend.repositories;
 import com.gerstox.projects.motostore_backend.entities.Recharge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RechargeRepository extends JpaRepository<Recharge, Integer> {}
+public interface RechargeRepository extends JpaRepository<Recharge, Long> {}
