+++ src/main/java/com/gerstox/projects/motostore_backend/repositories/StreamingRepository.java
@@ -3,4 +3,4 @@ package com.gerstox.projects.motostore_backend.repositories;
 import com.gerstox.projects.motostore_backend.entities.Streaming;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StreamingRepository extends JpaRepository<Streaming, Integer> {}
+public interface StreamingRepository extends JpaRepository<Streaming, Long> {}
