+++ src/main/java/com/gerstox/projects/motostore_backend/services/LicenseService.java
@@ -8,13 +8,13 @@ public interface LicenseService {
 
   public Page<License> findAll(Integer page, Integer elements);
 
-  public Optional<License> findById(Integer id);
+  public Optional<License> findById(Long id);
 
   public License save(License license);
 
-  public Optional<License> update(Integer id, License license);
+  public Optional<License> update(Long id, License license);
 
-  public Optional<License> delete(Integer id);
+  public Optional<License> delete(Long id);
 
-  public Boolean existsById(Integer id);
+  public Boolean existsById(Long id);
 }
